% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mrp_data.R
\name{clean_mrp_data}
\alias{clean_mrp_data}
\title{clean mrp data}
\usage{
clean_mrp_data(RawStr, country_iso, variable, period)
}
\arguments{
\item{RawStr}{raw table string scraped from market risk premia website}

\item{country_iso}{string containing 2iso codes}

\item{variable}{string indicating which variables to return}

\item{period}{string from t0 to t3}
}
\value{
df
}
\description{
clean mrp data
}
