% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crypto_data.R
\name{get_crypto_data}
\alias{get_crypto_data}
\title{Get crypto asset data from Binance}
\usage{
get_crypto_data(
  tickers,
  start_date,
  end_date,
  frequency = c("daily", "weekly", "monthly", "quarterly", "annual")
)
}
\arguments{
\item{tickers}{a character vector of crypto crosses, e.g. c('BTCBUSD', 'ETHUSDT')}

\item{start_date}{start date}

\item{end_date}{end date}

\item{frequency}{string indicating the frequency to convert data}
}
\value{
a tbl_df with columns ticker, date, variable and value
}
\description{
Get crypto asset data from Binance
}
