% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_equity_data.R
\name{get_equity_data}
\alias{get_equity_data}
\title{Wrapper function to companyDataScrapeR}
\usage{
get_equity_data(
  tickers,
  type,
  start_date = NULL,
  end_date = NULL,
  frequency = NULL
)
}
\arguments{
\item{tickers}{character vector of tickers where data is prefixed with the source, e.g. YAH-EXAMPLETICKER}

\item{type}{string specifying the type of data required - one of is, bs, cfs or price}

\item{start_date}{start date}

\item{end_date}{end date}

\item{frequency}{string indicating the frequency to return price data, e.g. 'monthly'}
}
\value{
object of class \code{tbl_df} with columns, date, ticker, type and nested company data
}
\description{
Wrapper function to companyDataScrapeR
}
