% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peresec_factor_data.R
\name{get_factors_by_weight}
\alias{get_factors_by_weight}
\title{get factors by weight}
\usage{
get_factors_by_weight(raw_factor_data, weight = c("equal", "market_cap"))
}
\arguments{
\item{raw_factor_data}{raw data retrieved from peresec website using tidyxl}

\item{weight}{string indicating whether to return equal or market cap weighted data}
}
\value{
tbl with cols  date, and factors e.g. Market, Size, Value, etc.
}
\description{
a cleaning function applied to the raw data retrieved from Peresec
}
