% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_macro_data.R
\name{get_macro_data}
\alias{get_macro_data}
\title{Import market data from Quandl and or AlphaVantage}
\usage{
get_macro_data(
  tickers,
  start_date,
  end_date,
  frequency = c("daily", "weekly", "monthly", "quarterly", "annual"),
  adjusted = TRUE
)
}
\arguments{
\item{tickers}{character vector of tickers where Quandl data is prefixed with Q- and AlphaVantage data is prefixed with AVST- for stocks and AVFX- for FX}

\item{start_date}{start date}

\item{end_date}{end date}

\item{frequency}{string indicating the frequency of data to retrieve - one of daily, weekly, monthly, quarterly or annual}

\item{adjusted}{logical indicating whether to retrieve adjusted prices where available}
}
\value{
a tbl_df with columns date, ticker, variable and value
}
\description{
Import market data from Quandl and or AlphaVantage
}
