% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peresec_factor_data.R
\name{get_peresec_factor_data}
\alias{get_peresec_factor_data}
\title{Get Factor Data for the South African Equity Market from Peresec}
\usage{
get_peresec_factor_data(
  reqd_file = c("long_only_alsi", "long_short_alsi", "quintile_alsi", "decile_alsi",
    "long_only_constrained", "long_short_constrained", "quintile_constrained",
    "decile_constrained"),
  reqd_factor = c("ff3", "ff5", "carhart4", "aqr6"),
  weight = c("equal", "market_cap"),
  download_directory = NULL
)
}
\arguments{
\item{reqd_file}{string indicating the required file e.g. long_short_alsi}

\item{reqd_factor}{string indicating the factors you require from the file e.g. ff3}

\item{weight}{string indicating whether to return equal or market cap weighted data}

\item{download_directory}{optional string indicating the destination directory for downloaded files - defaults to chromeDL temp dir}
}
\value{
wide tbl_df with date and factors
}
\description{
Get Factor Data for the South African Equity Market from Peresec
}
