% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quandl_data.R
\name{get_quandl_data}
\alias{get_quandl_data}
\title{get data from Quandl}
\usage{
get_quandl_data(tickers, start_date, end_date, frequency)
}
\arguments{
\item{tickers}{character vector of tickers prefixed with Q-}

\item{start_date}{start date}

\item{end_date}{end date}

\item{frequency}{string indicating the frequency - daily, weekly, monthly, quarterly, annual}
}
\value{
a tbl_df with cols ticker, date, variable and value
}
\description{
get data from Quandl
}
