% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mrp_data.R
\name{get_risk_premia_data}
\alias{get_risk_premia_data}
\title{get market risk premium data}
\usage{
get_risk_premia_data(country_iso = "all", variable = "all", period = "all")
}
\arguments{
\item{country_iso}{string indicating the countries to return using their 2iso codes - defaults to all available countries}

\item{variable}{string indicating which variables to return - defaults to all available variables}

\item{period}{string indicating which period to return - any of t0, t1, t2, t3 - defaults to all periods}
}
\value{
df with country_iso, date, variable and period
}
\description{
get market risk premium data
}
