% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_ticker_id.R
\name{get_ticker_names}
\alias{get_ticker_names}
\title{Wrapper function to companyDataScrapeR::get_ticker_id
This is a convenience function and shouldn't really need to be used. Names are now extracted in the meta data from calls to retrieve equity data}
\usage{
get_ticker_names(tickers)
}
\arguments{
\item{tickers}{a character vector of tickers, security names or ISINs}
}
\value{
a tbl_df with columns ticker, name and isin
}
\description{
Wrapper function to companyDataScrapeR::get_ticker_id
This is a convenience function and shouldn't really need to be used. Names are now extracted in the meta data from calls to retrieve equity data
}
