% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alphavantage_data.R
\name{pvt_retrieve_av_data}
\alias{pvt_retrieve_av_data}
\title{private function to retrieve daily data from alphavantage}
\usage{
pvt_retrieve_av_data(tickers, out_size, frequency)
}
\arguments{
\item{tickers}{a character vector of alphavantage tickers}

\item{out_size}{string indicating the output size - either full or compact}

\item{frequency}{string indicating the frequency - daily, weekly, monthly, quarterly, annual}
}
\value{
a list of dataframes
}
\description{
private function to retrieve daily data from alphavantage
}
