% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_ticker_id.R
\name{pvt_search_ticker_id}
\alias{pvt_search_ticker_id}
\title{private function to find individual ticker names}
\usage{
pvt_search_ticker_id(pjs_session, tkr)
}
\arguments{
\item{pjs_session}{remote driver object}

\item{tkr}{string indicating an ISIN, CUSIP, SEDOL, or security name or part thereof}
}
\value{
ticker name
}
\description{
private function to find individual ticker names
}
