% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peresec_factor_data.R
\name{scrape_peresec_factor_data}
\alias{scrape_peresec_factor_data}
\title{Access SA Factor Data Library from Peresec}
\usage{
scrape_peresec_factor_data(
  reqd_file,
  reqd_factor = c("ff3", "ff5", "carhart4", "aqr6", "peresec7lowvol",
    "peresec7lowbeta"),
  download_directory = NULL
)
}
\arguments{
\item{reqd_file}{string indicating the required file e.g. long_short_alsi}

\item{reqd_factor}{string indicating the factors you require from the file e.g. ff3}

\item{download_directory}{optional string indicating the download directory}
}
\value{
tbl_df with nested raw data returned from tidyxl
}
\description{
Access SA Factor Data Library from Peresec
}
