% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_ticker_id.R
\name{get_ticker_names}
\alias{get_ticker_names}
\title{find the full names of tickers given an ISIN, CUSIP, SEDOL or security name or part thereof}
\usage{
get_ticker_names(tickers)
}
\arguments{
\item{tickers}{a character vector of tickers}
}
\value{
a df of tickers and their full names
}
\description{
find the full names of tickers given an ISIN, CUSIP, SEDOL or security name or part thereof
}
